healthpilot/requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
asgiref==3.7.1
asttokens==2.0.5
async-timeout==3.0.1
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.2
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
channels==4.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==3.0.2
debugpy==1.5.1
decorator==5.1.1
Django==4.2.1
django-countries==7.5.1
django-channels==0.7.0
djangorestframework==3.14.0
docker-pycreds==0.4.0
docopt==0.6.2
drf-nested-routers==0.93.4
drf-yasg==1.21.5
flower==1.2.0
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.8
packaging==23.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.1
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
zope.interface==6.0